@@ -5,10 +5,11 @@ using System.Runtime.InteropServices;
 namespace FileSystem.Components
 {
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
-    public record struct FileHeader : IComparable<FileHeader>
+    public record struct FileHeader<T> : IComparable<FileHeader<T>>
+        where T : struct
     {
         public const int HeaderAddressOffset = sizeof(int); // Index
-        public static readonly int Size = Unsafe.SizeOf<FileHeader>();
+        public static readonly int Size = Unsafe.SizeOf<FileHeader<T>>();
 
         public int Index;
         public int NextHeaderAddress;
@@ -16,28 +17,29 @@ namespace FileSystem.Components
         public int CompressedSize;
         public int UncompressedSize;
         public uint ContentHash;
+        public T Metadata;
 
-        public int CompareTo(FileHeader other)
+        public int CompareTo(FileHeader<T> other)
         {
             return Index.CompareTo(other.Index);
         }
 
-        public static bool operator <(FileHeader left, FileHeader right)
+        public static bool operator <(FileHeader<T> left, FileHeader<T> right)
         {
             return left.CompareTo(right) < 0;
         }
 
-        public static bool operator <=(FileHeader left, FileHeader right)
+        public static bool operator <=(FileHeader<T> left, FileHeader<T> right)
         {
             return left.CompareTo(right) <= 0;
         }
 
-        public static bool operator >(FileHeader left, FileHeader right)
+        public static bool operator >(FileHeader<T> left, FileHeader<T> right)
         {
             return left.CompareTo(right) > 0;
         }
 
-        public static bool operator >=(FileHeader left, FileHeader right)
+        public static bool operator >=(FileHeader<T> left, FileHeader<T> right)
         {
             return left.CompareTo(right) >= 0;
         }
