@@ -1,10 +1,10 @@
-﻿namespace MapConverter
+﻿namespace FileConverter
 {
     public class Program
     {
         public static unsafe void Main(string[] args)
         {
-            CCMapConverter p = new(".\\", 4, 1448, 1448);
+            Converter p = new(".\\", 4, 1448, 1448);
             p.Convert("converted.bin");
         }
     }
