@@ -2,40 +2,68 @@
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
 using System;
+using UOClient.Utilities.Polyfills;
+using Matrix = Microsoft.Xna.Framework.Matrix;
+using Vector3 = Microsoft.Xna.Framework.Vector3;
 
 namespace UOClient
 {
     public sealed class IsometricCamera
     {
-        private Vector3 position;
-        private Vector3 target;
-        private float zoom;
+        private const int maxHeight = 128;
+        private const int nearPlane = maxHeight;
+        private const int farPlane = maxHeight * 3;
 
-        public Matrix WorldMatrix { get; private set; }
-        public Matrix ViewMatrix { get; private set; }
-        public Matrix ProjectionMatrix { get; private set; }
+        public const float HeightRateo = 7.5f;
+
+        private static readonly Vector3 positionFromOrigin = new Vector3(1, (float)Math.Sqrt(2), 1) * maxHeight;
+
+        private static readonly Vector3 left = new(-1, 0, 1);
+        private static readonly Vector3 right = new(1, 0, -1);
+        private static readonly Vector3 up = new(-1, 0, -1);
+        private static readonly Vector3 down = new(1, 0, 1);
+
+        private Vector3 target;
+        private readonly Matrix worldMatrix;
+        private readonly Matrix projectionMatrix;
+        private Matrix viewMatrix;
+        private Matrix scaleMatrix;
+        private Matrix worldViewMatrix;
+        private Matrix worldViewProjectionMatrix;
+        private Matrix invertedProjectionMatrix;
+
+        public ref readonly Matrix WorldMatrix => ref worldMatrix;
+        public ref readonly Matrix ViewMatrix => ref viewMatrix;
+        public ref readonly Matrix ProjectionMatrix => ref projectionMatrix;
+        public ref readonly Matrix ScaleMatrix => ref scaleMatrix;
+        public ref readonly Matrix WorldViewMatrix => ref worldViewMatrix;
+        public ref readonly Matrix WorldViewProjectionMatrix => ref worldViewProjectionMatrix;
 
         public Vector3 Target => target;
-        public Vector3 Position => position;
-        public float Zoom => zoom;
+        public int NearPlane => nearPlane;
+        public int FarPlane => farPlane;
+        public float Zoom { get; private set; }
 
         public IsometricCamera()
         {
-            float val = (float)Math.Cos(Math.PI / 4);
+            Zoom = 1;
 
-            target = new(249, 0, 1026);
-            position = (new Vector3(val, val * 10, val) * 127) + target;
-            zoom = 1;
+            target = Vector3.Zero; //new(712, 0, 1367); //new(835, 0, 904);
+            scaleMatrix = Matrix.CreateScale(Zoom, Zoom, 1);
 
-            WorldMatrix = Matrix.Identity;
-            ViewMatrix = Matrix.CreateLookAt(position, target, Vector3.UnitY);
+            worldMatrix = Matrix.CreateScale(1, 1 / HeightRateo, 1);
 
-            ProjectionMatrix = Matrix.CreateOrthographic(20, 20, 0, 3000.0f)
-                * Matrix.CreateScale(1, 1.5f, 1);
+            projectionMatrix = Matrix.CreateTranslation(-0.5f, -0.5f, 0)
+                * Matrix.CreateOrthographic(20, 20, nearPlane, farPlane)
+                * Matrix.CreateScale(1, (float)Math.Sqrt(2), 1);
+
+            UpdateMatrices();
         }
 
-        public void HandleKeyboardInput()
+        public bool HandleKeyboardInput()
         {
+            bool modified = false;
+
             KeyboardState keyboard = Keyboard.GetState();
 
             bool up = keyboard.IsKeyDown(Keys.Up);
@@ -43,45 +71,54 @@ namespace UOClient
             bool right = keyboard.IsKeyDown(Keys.Right);
             bool down = keyboard.IsKeyDown(Keys.Down);
 
-            if (keyboard.IsKeyDown(Keys.OemPlus))
-                zoom += .01f;
-            else if (keyboard.IsKeyDown(Keys.OemMinus))
-                zoom -= .01f;
+            if (keyboard.IsKeyDown(Keys.OemPlus) && Zoom < 1)
+                UpdateScale(.01f);
+
+            else if (keyboard.IsKeyDown(Keys.OemMinus) && Zoom > 0.1f)
+                UpdateScale(-.01f);
 
             if (up)
-            {
-                position.X -= 1;
-                position.Z -= 1;
-                target.X -= 1;
-                target.Z -= 1;
-            }
+                UpdatePosition(IsometricCamera.up);
 
             if (right)
-            {
-                position.X += 1;
-                position.Z -= 1;
-                target.X += 1;
-                target.Z -= 1;
-            }
+                UpdatePosition(IsometricCamera.right);
 
             if (left)
-            {
-                position.X -= 1;
-                position.Z += 1;
-                target.X -= 1;
-                target.Z += 1;
-            }
+                UpdatePosition(IsometricCamera.left);
 
             if (down)
+                UpdatePosition(IsometricCamera.down);
+
+            if (modified)
+                UpdateMatrices();
+
+            return modified;
+
+            void UpdatePosition(Vector3 vector)
             {
-                position.X += 1;
-                position.Z += 1;
-                target.X += 1;
-                target.Z += 1;
+                target += vector;
+                modified = true;
             }
 
-            ViewMatrix = Matrix.CreateLookAt(position, target, Vector3.UnitY)
-                * Matrix.CreateScale(zoom, zoom, 1);
+            void UpdateScale(float increment)
+            {
+                Zoom += increment;
+                scaleMatrix.M11 = Zoom;
+                scaleMatrix.M22 = Zoom;
+
+                modified = true;
+            }
+        }
+
+        private void UpdateMatrices()
+        {
+            viewMatrix = Matrix.CreateLookAt(target + positionFromOrigin, target, Vector3.Up);
+            MatrixUtilities.Multiply(in viewMatrix, in scaleMatrix, out viewMatrix);
+
+            MatrixUtilities.Multiply(in worldMatrix, in viewMatrix, out worldViewMatrix);
+            MatrixUtilities.Multiply(in worldViewMatrix, in projectionMatrix, out worldViewProjectionMatrix);
+
+            MatrixUtilities.Invert(in worldViewProjectionMatrix, out invertedProjectionMatrix);
         }
 
         public void Test(GraphicsDevice device)
